api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
httpx==0.28.0
pydantic==2.10.0
pydantic-settings==2.7.0
GEOparse==2.0.4
scanpy==1.10.4
anndata==0.11.1
numpy==1.26.4
pandas==2.2.3
scipy==1.14.1
scikit-learn==1.5.2
plotly==5.24.1
redis==5.2.1
python-jose[cryptography]==3.3.0
supabase==2.11.0
gunicorn==23.0.0
statsmodels==0.14.4
